database.read: when snapshot val does not exists
When trying to `read` a value in a path that doesn't exist,
the current behavior is that it throws an exception because of result.val is not a function.
I'd like that it returns a null value instead. 

A use case, in particular, Imagine a user can `like` a `post`
```
user: {
  u1: { name: 'alice' },
  u2: { name: 'bob' }
},
post: {
  t1 : { title: 'something' },
  t2 : { title: 'something else ' }
},
post-likes: {
  t2: {
      u1: { timestamp: 1518665595337 }
  }
}
```
when reading if a user liked a post I'd need to 
If I want to know if bob liked the second post then I need to read 
```
database.read(`post-likes/t2/u2`)  
```
because t2.u2 doesn't exist it will throw an uncaught exception. 
alternatively, if you don't like the idea of returning null, 
it could throw an exception with an informative message or code. 
something like 
```
  const exists = yield call(result, exists)
  if (!exists) throw new Error('path does not exsits') 
```